feat(portfolio): implement total staked calculation for projects (#6122)
# Motivation

The `Portfolio` page features a card that displays the total assets
owned by the user. In this second iteration, Neurons are used to
calculate wealth. A final PR will provide both Tokens and Neurons from
the `Portfolio` route to the page.

The same design discrepancies mentioned in #6121 remain.

# Changes

- Exposes the `tableProjects` prop on the `Portfolio` page to receive
the list of all projects and the total amount staked.
- Adds logic to display the `NoNeuronsCard` when the neurons balance is
zero.

# Tests

- Unit test to verify the presence of the `NoNeuronsCard` based on the
neurons balance.
- Unit test to confirm that a primary action appears in the
`NoNeuronsCard` when the token balance exceeds zero.
- Unit test to ensure that the `UsdValueBanner` displays the correct
value based on the token balance.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

Prev. PR: #6121