Recover from interrupted canister creation in frontend (#6132)
# Motivation

Currently, the NNS dapp backend canister monitors every transaction on
the ICP ledger in order to catch canister creation funding transactions
for which the notify+attach process might have been interrupted.
We want to remove this logic from the backend and perform it in the
frontend in order to simplify the responsibilities of the backend
canister.

In this PR we check if loaded transactions contain any canister creation
funding transactions for which there is no corresponding canister.
This is done if the user visits the wallet page after visiting the
canisters page.
Since it should be rare for canister creation to be interrupted, we
assume that when it happens, a user will look around for their canister
and eventually trigger the recovery process.

# Changes

1. In `NnsWallet`, if the canisters are loaded, check if we need to
notify for interrupted canister creation.

# Tests

1. Unit tests added.
2. Tested manually with a backend canister where the fallback process is
disabled.
3. Tested manually with a backend canister that still has the fallback
process as well to make sure this doesn't break anything until it's
remove from the backend.

# Todos

- [x] Add entry to changelog (if necessary).