NNS1-3504: Hide Cycle Transfer Station from Voting page (#6144)
# Motivation

The Cycles-Transfer-Station SNS was renamed to `----` and made unusable
by impossible proposal fees and transaction fees.

To let people see their former balance and neurons we are keeping it on
the Tokens and Staking pages for now but we're hiding it from the Voting
page.

# Changes

1. Move `CYCLES_TRANSFER_STATION_ROOT_CANISTER_ID` from
`frontend/src/lib/stores/sns-aggregator.store.ts` to
`frontend/src/lib/constants/canister-ids.constants.ts`.
2. Hide the former "Cycles-Transfer-Station" SNS from the
`SelectUniverseList` which (since the tokens/neurons tables UIs) is only
used on the Voting page.

# Tests

1. Simplify test setup by using `setSnsProjects` instead of mocking
`snsProjectsCommittedStore`.
2. Add a unit test with a project with the same root canister ID as
"Cycles-Transfer-Station".
3. Tested manually against mainnet.

# Todos

- [x] Add entry to changelog (if necessary).