feat(portfolio): implement TokensCard component (#6157)
# Motivation

The `TokensCard` component on the Portfolio page displays the total USD
amount of tokens and a list of up to four tokens. If there are fewer
than three tokens, it shows a section with additional information.

This component will receive data from the Portfolio page. The page will
take the user's token list and convert it into a list of four relevant
tokens because other components on the page also require this
information.

A future PR will enhance the styling of the subcomponents.

Related PRs:
- Get top tokens #6153

# Changes

- Adds the `TokensCard` component.

# Tests

- Unit tests for the component  
-  New page object for the component 
- Manually tested in [E2E
environment](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/)

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary