feat(portfolio): implement token cards relationship management (#6183)
# Motivation

There is a relationship between the `HeldTokensCard` and the
`StakedTokensCard`. Based on this relationship, we display or hide an
additional row with a static message to achieve visual balance between
the heights of the cards.
The logic for this can be found in #6181. 

Live demo can be found
[here](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network)

# Changes

- Updates `HeldTokensCard` to receive the number of staked tokens and
determine whether to display the `InfoRow`.
- Updates the Portfolio page to provide the number of staked tokens to
`HeldTokens`.
- Updates `StakedTokensCard` to utilize a new utility function for
deciding whether to display the `InfoRow`.
- Includes some minor CSS changes.

# Tests

- Update tests based on the new logic

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary