feat(portfolio): add a stake loading state field to ProjectTable (#6201)
# Motivation

We want users to have cue that their tokens are loading. The `UserToken`
value already has this functionality, represented by `UserTokenLoading`.
The goal is to implement a similar feature for the `ProjectTable`.

In this initial iteration, I'm adding a new property instead of changing
the existing `stake` property to minimize the impacted area.

Relates to #5199

# Changes

- Added a new property `isStakeLoading`, which is `true` while the
neurons have not yet loaded.

# Tests

- Updated tests to check the new value when neurons have not loaded yet

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary