feat(portfolio):  implements SkeletonTokensCard component (#6208)
# Motivation

Some users on certain networks may experience high loading times for
their assets. We want to display loading states during this period.

This first PR introduces a new skeleton component specifically for the
Portfolio page. We already have a `SkeletonCard`, but since this is the
entry point for the application, we want it to resemble the final cards
as closely as possible.


https://github.com/user-attachments/assets/3c6126b2-8b63-42c0-9437-4de68e78310c

# Changes

- New skeleton component for the Portfolio page.

# Tests

- Not necessary, as this is a pure UI component. Future PRs will
integrate it into the Portfolio page, where we will test its visibility
when needed.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary