FOLLOW-1337: Disable FORCE_CALL_STRATEGY for 30% of sessions (#6205)
# Motivation

We want to re-enable the update calls that are disabled via
FORCE_CALL_STRATEGY.
Because this may result in higher subnet load, we have first rolled it
out to 10% of sessions in [this
PR](https://github.com/dfinity/nns-dapp/pull/5593).
Just before that rollout, subnets got overloaded for unrelated reasons,
but are now back to normal.
So we will proceed to 30% to see if it has any effect on subnet load.

# Changes

1. Initialize `FORCE_CALL_STRATEGY` on app load, to `"query"` 70% of the
time and to `undefined` 30% of the time.

# Tests

Not tested.

# Todos

- [x] Add entry to changelog (if necessary).