Migrate LosingRewardsBanner to voting power economics (#6238)
# Motivation

Currently, the voting power parameters are hardcoded as half a year and
one month. However, actual values are already being retrieved from the
API (they are the same for now but may be updated in the future). The
general idea is that if the parameters are unavailable, we consider the
neuron active.

In this PR, we migrate `LosingRewardsBanner` from constants to the
voting power economics store.

# Changes

Switch from constant based to VP economics stores and utilities.
Do not show the component without voting power economics available (it
is loaded on dapp initialisation).

# Tests

- Updated.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.