Migrate getNeuronTags to voting power economics (#6241)
# Motivation

Currently, the voting power parameters are hardcoded as half a year and
one month. However, actual values are already being retrieved from the
API (they are the same for now but may be updated in the future). The
general idea is that if the parameters are unavailable, we consider the
neuron active.

In this PR, we migrate the getNeuronTags function (and its callers) to
the voting power economics utilities.

# Changes

- Switch to ...VPE utils in getNeuronTags
([changes](https://github.com/dfinity/nns-dapp/pull/6241/files#diff-b570daeb2a65094f5bd1d1b823bcf3025de7af68555903e0c2061d2f8b029e8fR555-R581)).
- Update all callers to provide startReducingVotingPowerAfterSeconds
argument.

# Tests

- Updated.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.