Switch to shouldDisplayRewardLossNotificationVPE (#6249)
# Motivation

Currently, the voting power parameters are hardcoded as half a year and
one month. However, actual values are already being retrieved from the
API (they are the same for now but may be updated in the future). The
general idea is that if the parameters are unavailable, we consider the
neuron active.

In this PR, we replace the constants based
`shouldDisplayRewardLossNotification` function with
`shouldDisplayRewardLossNotificationVPE`.

# Changes

- Replace in `soonLosingRewardNeuronsStore`.
- Replace in NnsNeuronDetail for ConfirmFollowingBanner visibility.

# Tests

- Updated. The NnsNeuronDetail was already covered because of the partly
duplicated logic in the ConfirmFollowingBanner
[here](https://github.com/dfinity/nns-dapp/pull/6236/files#diff-1cfb1fc9513a792bdab9ae3014a9012736bbf26f57001b3c1adcbddb51d769feR156-R172).
- Update NnsNeuronsMissingRewardsBadge.spec, because the component
depends on soonLosingRewardNeuronsStore.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.