Fix preventing global fetch in tests (#6253)
# Motivation

We have code to prevent tests from doing `fetch`. When `fetch` happens
during tests it typically means that they should have mocked an API call
that's being used.

But this code wasn't working because it uses mocking and
`vi.restoreAllMocks()` was being called after the setup was done.

# Changes

1. Mock `global.fetch` in `beforeEach` instead of directly in
`vitest.setup.ts`.

# Tests

1. Add missing API mocking in tests that started failing.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary