Remove last_ledger_sync_timestamp_nanos from AccountsStore (#6268)
# Motivation

The nns-dapp canister no longer processes ICP transactions so we don't
need to keep track of the last time we processed transactions.

# Changes

1. Remove field `last_ledger_sync_timestamp_nanos` from `AccountsStore`.
2. When encoding the `AccountsStore` to stable memory, write `0u64` in
the slot of `last_ledger_sync_timestamp_nanos`.
3. Ignore the slot of `last_ledger_sync_timestamp_nanos` when decoding
the `AccountsStore` from stable memory.

# Tests

This was was already not used anymore so no tests needed to be changed.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary