Import token by url (#6257)
# Motivation

To simplify the token import process, the user can import a custom CK
token by clicking a link, and the import token form will be pre-filled
with the parameters from the URL (e.g.
//nns.ic0.app/tokens/?import-ledger-id=m7h5t-euaaa-aaaaa-qabia-cai&import-index-id=myg3h-jmaaa-aaaaa-qabiq-cai).
Having the ledger canister ID is sufficient to add an imported token.

In this PR, we fill in the form and automatically start validation.

1. Validate the canister IDs provided in the URL.
2. If the ID(s) are valid, fill in the form.
3. Start validation.

**Out of scope:**
- Additional handling of same tokens import.
- Sign-out flow.

# Changes

- Add logic to start the import token process by URL.

# Tests

- Added.
- Tested locally.


https://github.com/user-attachments/assets/22143018-db80-45eb-a4ab-336dfd12de89

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.