Import token sing out modal (#6276)
# Motivation

To simplify the token import process, the user can import a custom icrc1
token by clicking a link, and the import token form will be pre-filled
with the parameters from the URL (e.g.
//nns.ic0.app/tokens/?import-ledger-id=m7h5t-euaaa-aaaaa-qabia-cai&import-index-id=myg3h-jmaaa-aaaaa-qabiq-cai).
Having the ledger canister ID is sufficient to add an imported token.

In this PR, we display the import token modal for signed-out users.

# Changes

- Display the modal on the SignOutTokens page with a sign-in button.

# Tests

- Added.
- Tested manually as well
([demo](https://qsgjb-riaaa-aaaaa-aaaga-cai.mstr-ingress.devenv.dfinity.network/accounts/?u=qsgjb-riaaa-aaaaa-aaaga-cai))

![image](https://github.com/user-attachments/assets/2eacf75f-534c-47ae-bb37-0d6cb3a54ca5)

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary