refactor(navigation): enhance referrer path tracking with list (#6274)
# Motivation

The Accounts page features a custom back button that currently redirects
users to the Tokens page. This is no longer always accurate, as users
can now access this page from the Portfolio page.

Since users can also navigate from the Accounts page to the Wallet page,
we cannot rely on a local solution. The Accounts layout will be
unmounted, resulting in the loss of this information.

This first PR modifies the existing solution used by the Proposal page
to redirect back to either the Proposals or Launchpad by tracking the
referrer. To gain more insight than just the last jump, the current
single value has been changed to a list of referrers.

A follow-up PR will use this updated store in the Accounts layout to
track how users enter the page and determine where to navigate when they
click the custom back button.

Note: We do not want to rely on the browser's history stack. If a user
directly enters the Accounts page, we want to navigate them back to the
Tokens page as we currently do.

# Changes

- Updates the `referrerPathStore` to maintain a list of referrers.

# Tests

- Updates unit tests affected by the change.  
-  Adds new unit tests for the updated store.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary