NNS1-3548: Add compare functions for projects table (#6283)
# Motivation

We want to allow the user to change the order in the projects table by
clicking on the table headers, similar to the neurons table.
When clicking the "Nervous Systems" header, the projects should be
ordered alphabetically by name.
When clicking the "Neurons" header, the projects should be ordered
number of neurons.
When clicking the "Stake" header, the projects should be ordered by:
1. USD value of the stake
2. Projects without exchange rate but positive number of neurons before
projects with zero neurons.
3. Within projects with the same (probably 0) USD stake, ICP should be
sorted before others.

This PR adds the sorting functions.
It does not add the functionality in the table.

# Changes

1. Add necessary compare functions.

# Tests

1. Unit tests added.
2. Tested manually (with additional changes) at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/tokens/

# Todos

- [ ] Add entry to changelog (if necessary).
not yet