NNS1-3548: Allow disabling table sorting on mobile (#6294)
# Motivation

We want to allow sorting the tokens table.
On desktop, sorting the table is done by clicking on the table column
header.
But on mobile, there is an icon in the top-right to open a sorting
modal.
However, the tokens table already has an icon in the top right.
Until we have a better solution for that, we've decided to just not
allow sorting the tokens table on mobile so that we can already allow
sorting the tokens table on desktop.

This PR adds a prop to the `ResponsiveTable` component to control
whether the sorting icon should be displayed on mobile.

# Changes

1. Add `disableMobileSorting` prop to `ResponsiveTable` to hide the
sorting icon button.

# Tests

1. Unit test added.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet